Rework the driver package interface.
Summary:
Separate the callbacks for setup, teardown, and error reporting from the
delivery of output. Pull out the handling of unset callbacks so that the main
flow of Run is less disrupted. Rename/rework the IO type as "Context".

Tests updated, etc.

Fix driver setup in the analysis_driver binary.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1453