Use tapps to refer to polymorphic functions where possible
Summary:
This change partially addresses T229.

Member functions of class templates are currently
type-erased (and were before this CL in certain contexts).
We'll need to record the whole type context in their
tapps if we want to distinguish between various
instantiations (since we no longer get to point at
the subgraph where we'd otherwise be able to get the
information).

Reviewers: fromberger, #core_team

Reviewed By: fromberger, #core_team

Subscribers: #core_team

Maniphest Tasks: T229

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1470