Go: Handle unresolved identifiers more gracefully.
Summary:
In some cases, a package that is structurally valid may be missing some
dependencies leading to identifiers that cannot be resolved. Prior to this
change, that could lead to a panic because we were assuming a typechecked
package must have all its definitions.

This change makes the indexer more forgiving: In cases where a definition does
not exist we will now log and continue.

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1475