Go: Add support for MarkedSource "code" facts.
Summary:
This change adds minimal support to the Go indexer for emitting code facts with
MarkedSource messages. Right now the implementation is just a dump of the text
format generated by the go/types package, without any further structure. T232
tracks generating a more complete representation.

This also adds a flag to the indexer tool to enable code facts to be emitted
when testing for them, and a smoke test to verify that the structure we are
emitting gets properly plumbed.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1479