Go: Mitigate a crash resulting from tag mismatch.
Summary:
If the indexer gets a compilation whose source list contains files with
incompatible build tags,[1] it is possible to drive the indexer into a state where
it has competing incompatible declarations for what should be a single type or
function.

To fully fix this, we will need to be smarter about handling build tags; if the
extractor can't do this sensibly (which apparently it doesn't now), we may have
to teach the indexer to handle it.

For the moment, just stifle the crash.

[1] http://godoc.org/go/build#hdr-Build_Constraints

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1491