Go: Filter source inputs based on build tags.
Summary:
The structure of a Go package allows build tags that control whether each file
is to be passed to the compiler (see https://godoc.org/go/build).

Previously we weren't handling these, since Bazel (our primary build tool) does
not use them. The native Go tool does use them, though, so for packages that
come in from outside we need to cope with these. Otherwise we may wind up
giving the compiler multiple files with different definitions of the same
entities, generating collisions (build tags are often used as a trivial kind of
conditional import).

We address this by abusing the go/build package's Context type, where the logic
for checking tags is vaguely exposed. This is a bit of a hack, but go/build is
mostly protected by the Go 1 guarantee so it should be fairly safe.

Related:
- Add a Go-specific compilation details message.
- Make the Bazel and go tool extractors use the GoDetails message.

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: salguarnieri, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1492