Attach documentation for children when requested
Summary:
This diff adds an option (include_children) to DocumentationRequest.
This option adds the documentation for (moral) children of the
request's ticket for display in some structured documentation view
(like Javadoc). Clients can't implement this completely on their
own without use of the GraphService because it relies on snooping
around the `childof` subgraph.

Reviewers: schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1487