Add a final_result field to the AnalysisOutput message.
Summary:
We would like to allow analyzers to report status information back to the
driver. However, the protobuf service definition does not provide any clear
mechanism for doing so; the propagation of status information depends on the
RPC transport used.

This change provides a way for analyzers to report status in-band: The analyer
MAY set the final_result field of its last AnalysisOutput message to the
driver, in lieu of an output value. To ensure orderly operation, the analyzer
may not send any further outputs after that has occurred, and the driver must
enforce this.

- Documentation updated.
- Update generated protobuf code for Go.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1504