Don't insert annotations for nullability attributes
Summary:
Objective-C allows the user of the nullability attribute. In source,
this appears to the left of types (nullable Foo *). When we see it in
clang though, it has already been translated into the more standard
_Nullable attribute that appears on the right of types (Foo *
_Nullable).

Two things cause us trouble:
1) The source range for "nullable" is backwards. The start is at the "e"
in nullable and the end is at "n".
2) When we use logic to find a better range, we are exploring the
transformed source, so we are unable to find a reasonable range for the
token because it doesn't exist in the source, as we know it, any longer.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1505