Ported kindex file generator to GO.
Summary: This was necessary due to the fact that the original script worked with TextFormat protos, running into a limitation when a multi-line input file was encountered. Namely, it would strip newlines from the input file, which causes downstream failures. After investing some time into debugging it, and consulting with fromberger@, we decided it would be best to port the tool to Go so that it could work directly with binary protos to alleviate the problem.

Reviewers: fromberger, #core_team

Reviewed By: fromberger, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1511