Go indexer: Handle "anonymous" struct fields.
Summary:
Embedded unnamed struct field declarations create fields that can be referenced
by the base name of the declared type. Handle these by considering the base
name a binding site for the field. These are "slightly special" in that we need
to take care to avoid emitting duplicate anchors over the names of the types.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1517