Handle basic case for Attributed Type
Summary:
Attributed types (like `Data * _Nullable varname`) should link back to
the modified type. The attributed version is not really a new type, it
is just a specification of how this instantiation of the type should
behave. In the future, we might want to record facts about these attributes.

Test Plan: Added objc attribute test that would previously fail.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1521