Add javac to third_party
Summary:
Previously kythe was depending on the version of javac that ships with
Bazel. This change makes it easier to upgrade the javac version
independently from Bazel releases. The javac artifact that was added is
the same version that shipped with Bazel 4.5.

This change also updates BUILD files to use the new dependency, and
fixes a few places where javac or javac-based tools weren't using the
-Xbootclasspath/p: flag to prepend javac to the bootclasspath.

Test Plan: unit tests

Reviewers: fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1540