Don't use body location for macro marked source
Summary:
When we record marked source for a macro, we care about the macro's text
locations. In the old code, if the macro defined a function, we would set the
end source location to the start of the function body. This is a problem
because that end location doesn't really mean anything for the macro
text, it may be *before* the start of the macro text, and it may be in a
different file.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Maniphest Tasks: T241

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1556