Go indexer: Expand the treatment of MarkedSource.
Summary:
For variable, type, and function declarations, provide enough context for the
renderer to extract identifiers and basic type information.

Types are somewhat abbreviated, as the interior structure of struct and
interface declarations is generally not that interesting in the UI without the
code present -- we may want to revisit that later if we start doing more
elaborate things with the structure.

Add test cases for code facts.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1539