Add -scoped flag to the extract_kindex tool.
Summary:
A common case for matching source inputs is to restrict attention to only those
files governed by the rule itself. We can't easily do this without Bazel in
harness, but we can use the package's path as a proxy for that.

When -scoped is set, in addition to matching any -source regexp, a path must
also contain the package's directory path as a substring. This may pick up
other files in the same package, but averts the much worse case of capturing
files in other packages.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1573