typescript: allow caller to provide readFile() function
Summary:
The caller provides a loaded ts.Program that contains all the source
files, but unfortunately for the purposes of computing UTF-8 file
offsets we still need access to the raw file bytes.

We implemented this by calls to fs.readFileSync.  In environments
where the indexer doesn't have access to the file system we want
the caller to be able to provide the file-reading function.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1576