testutil: Add support for protobuf messages to DeepEqual.
Summary:
It is not generally safe to compare protobuf messages with the normal equality
tests (==, reflect.DeepEqual), since their internal structures may differ even
if the messages represented are equivalent. For proto3 syntax it was previously
not a problem in practice, but more recent versions of the code generator are
starting to include changes that touch this constraint.

Add a test to the DeepEqual function to explicitly check for protobuf messages
and to test them for equality before doing further reflective checks. This
ensures we will only report a difference for protobuf messages if there really
is a semantic difference.

Reviewers: schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1582