typescript: handle possible symbol lookup fail
Summary:
The TypeScript API can return 'undefined' from getSymbolAtLocation,
but the types don't capture that because they haven't yet migrated
to being null-safe.  Instead, wrap that function with one with the
proper return type, then make the compiler happy again by
null-checking the return value each time it's used.

This fixes indexing of lib.es6 (the TS standard library covering
ES6).  Before it was crashing when we attempted to get a symbol
name from a class property that was a computed variable.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1592