Go indexer: Mark new type declarations as records.
Summary:
Previously type definitions defaulted to the node kind "tapp", which was an
incorrect placeholder. For type definitions whose underlying type is not
already a struct or interface, denote these instead as a "record" node with the
new subkind "type".

Nominally, the graph model is of a record structure having a single unexported
field of the underlying type, which isn't actually how it is represented in Go,
but is close enough to give sensible graph results.

Schema: Add Go-specific rules for type definitions.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1594