Add metadata plumbing to the Go resolver.
Summary:
This change adds a hook to the resolver settings that allows linkage metadata
inputs to be discovered and loaded from among the required inputs of the
compilation.

Nothing is yet done with these data if they are found, and the default is the
current behaviour in which nothing is recognized as linkage metadata.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1605