Go indexer: Handle field declarations in anonymous types.
Summary:
It is legal in Go to declare the type of a variable or parameter as having an
anonymous struct type. When this occurs, references to the fields of the type
would be improved by having something to point to, e.g.

     s := struct{ T int }{
       T: 25, // a reference to T above
     }
     s.T = 30 // a reference to T three lines above

This change adds that support, and a test thereupon.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1624