Go indexer: Fully-qualify method owners in the package being indexed.
Summary:
When a package is imported, the exported methods on its types will be called as
qualified by their import path, not just the package name. But for top-level
types, we weren't properly qualifying the vname signatures for their methods at
the point of definition.

This change fixes that by ensuring that when we do construct the name of such a
method, it is qualified by its enclosing package's vname, even if that package
is local. Mainly, the change works by overriding the default string rendering
behaviour from the go/types package, whose default behaviour isn't sufficient
for this purpose.

Tests updated. Default test packages to simpler names and lay out extra data
files relative to the source root.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1632