Fix forward declaration completion edges for ObjC
Summary:
It looks like the right way to determine if something is an @interface
or an @class is to check if it is the definition. This does not mean
that it is the definition of the class, it means it is the definition of
the @interface. We still consider the @implementation to be the
definition of the class.

Test Plan: new test that previously failed.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1636