Java indexer: remove -Xlint and -Werror options
Summary:
Apparently JavacTool does not correctly handle some -Xlint options when
determining arity.  This can lead to proceeding options being omitted
erroneously and javac crashing when later invoked with the processed
arguments.

For now, remove all -Xlint and -Werror options since we do not use them
during the analysis.  Since we may change this logic in the future and
miss this tricky issue, I've also added a test that asserts the
erroneous JavacTool behavior to check for changes in future updates.

Reviewers: fromberger, #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: jrtom, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1638