Go indexer: Add links to godoc for each indexed package.
Summary:
Besides comments, package documentation can also be reached via a godoc server.
Probably the best-supported public version of this is at http://godoc.org/, but
it is also possible to run local mirrors. This change permits the Go indexer to
emit a doc/uri fact for each indexed package pointing to a godoc instance, if
one is available.

To do this:
- Add the doc/uri fact type to the schema package.
- Allow specifying a base URI for godoc links.
- Add --docbase flag to the go_indexer binary (for tests).
- Update tests for doc/uri facts.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1641