Go extractor: Updates for "improvements" to the upstream Go rules.
Summary:
The latest incarnation of the Bazel rules for Go includes some additional shell
nonsense as part of the compile command. The result is that we no longer get
the list of source inputs as part of the compiler command directly, but only as
after shell evaluation.

We can't replay the response file directly, as the extra action does't have
write access to that part of the sandbox. So for now, just work backward as we
did for the symlink forest. This is a little grody, but does work.

Reviewers: schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: schroederc, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1642