Go indexer: Simplify and improve MarkedSource representation.
Summary:
This commit flattens out the marked source representation for various Go
entities, simplifying the process of extracting identifiers and qualified
identifiers for presentation. The overall layout is mostly the same, the main
change here is to collapse some of the weird spacing and labelling I had done
in the original implementation.

Motto: MarkedSource is less complicated than you think. :)

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: zarko, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1644