typescript: properly handle rootDirs
Summary:
Under tools like bazel, we have multiple directory hierarchies that should
be treated as "the same" for semantic name purposes -- a file in the source
tree can import "./foo" even if foo is generated and live in a parallel
bazel-bin/... directory.

This semantic mapping is handled by TypeScript with the rootDirs option.
When mapping an on-disk path back to its semantic name, we must consult
rootDirs.  This generalizes the use of sourceRoot.

Reviewers: alexeagle, #core_team, fromberger

Reviewed By: alexeagle, #core_team, fromberger

Subscribers: alexeagle, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1672