typescript: index arrow functions
Summary:
Treat arrow functions as function-like objects for the purposes of
traversal, so that we index their declared parameters as variable
declarations.

Treat arrow functions as new scopes for the purposes of generating
signatures, so that an outer "x" and an "x" declared within the
arrow function have different signatures.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1684