Go indexer: Fix an incorrect marked source instance.
Summary:
Left over from a previous iteration, the root of the marked source was carrying
a post-child token to denote the separator between the context and its ident.
Now that the context deals with its own separators, this is unnecessary and in
fact wrong: When the context is injected, it results in the extra separator I
had thought was spurious. (It wasn't).

- Test restored.
- Fix up code_interface_test.
- Fix up code_method_test.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1686