typescript: properly resolve imports
Summary:
An import creates an alias into another module.
Rather than using the TS API to look up that name in the module's
exports, use the TS API that just resolves the alias.

This removes a bunch of hacks and simplifies some code.  And more
importantly, it properly handles the case where there's a chain
of aliases (e.g. a imports b, while b is a reexport of c).

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: fromberger, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1689