Work around an apparent struct-field bug in go/types.
Summary:
It appears that the behaviour of NumMethods/Field on a *types.Struct value does
not include the fields of embedded "anonymous" type, although the comments
suggest it should. In particular, the count is off and may induce a panic in
otherwise well-formed code when the initializer count differs from the number
of fields expected.

This is a workarond to keep the indexer from crashing while we sort out a
better solution.

Reviewers: #core_team, jrtom

Reviewed By: #core_team, jrtom

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1722