Correctly handle members of Java array classes.
Summary:
According to the language spec, all arrays with a given component type T have
the same type, T[]. This type is not a class, but it has an associated Class object.
(see https://docs.oracle.com/javase/specs/jls/se8/html/jls-10.html#jls-10.8)

In practice, the compiler generates a class `Array`, which is a holder for the array
members `clone()` and `length`. This single class is used behind the scenes to implement
those two members on behalf of all array types. This is merely an implementation detail, but
nonetheless appears in compiler data structures, and so Kythe ends up acting on it
when indexing references those members, instead of on the true array type.

Previously we also failed to recognize this `Array` class as a JDK class, and ended up
falling through to emitting a NAME node for `Array.clone()`, instead of a FUNCTION node.

This change:
- Provides the true array type to locations that need it.
- Ensures that we detect `Array` as a JDK class.

As a result, we now emit a FUNCTION node with the signature (and corresponding MarkedSource
entries) `int[].clone()`, instead of a NAME node with the name `Array.clone()`.

Reviewers: zarko, schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1725