extractschema: Capture notes on naming conventions.
Summary:
Add a field to the schema to capture default vname rules, and a field to the
node to capture specific overrides to the conventions. Include references from
these in the set of related kinds.

Make note of any field-specific vname rules for any node kind that defines them
beyond the core set.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1743