Add Go To Definition to language server
Summary: This change adds definition lookups to the server. Overall it's a small change because most of the work was done with references.

Test Plan: I'd like to find a good way to do end-to-end testing but that's gonna require a good chunk of infrastructure. If you have any ideas, let me know

Reviewers: fromberger, zarko, schroederc

Reviewed By: zarko, schroederc

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1773