Add a simple Java analyzer plugin framework
Summary:
Any number of Plugins can register themselves with the
KytheJavacAnalyzer to be run over each JCCompilationUnit.  The Plugin is
given an instance of KytheEntrySets to emit data.  It is also given a
mapping from JCTree to the VName of the node emitted by the Java
analyzer so that it can emit edges into the Kythe graph.

Reviewers: fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1809