Bazel extractor: Separate spawn-specific logic from extraction.
Summary:
In order to support some of the "native" Bazel rules, separate the part of the
extractor that handles the relatively-generic parts of the rule plumbing, viz.
inputs, outputs, arguments, target, and environment, into their own type.  The
extraction machinery now operates on that type, allowing the caller to handle
other rule types on a case-by-case basis.

The existing tools have been updated to preserve the spawn-specific logic.  We
might want to pull those apart further in the future.

Update unit tests.

Update extractor uses for the switch to ActionInfo.

Reviewers: shahms, #core_team

Reviewed By: shahms, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1790