Implement diffing for dirty states
Summary:
This change introduces diffing dirty states to allow for realigning of
references. In order to do this, the server maintains a copy of the original
source text and the dirty text. On requests for references and definitions,
the two files are diffed, and all references are given a mapped location.
Mapping happens again for all references located in open files so the reference
site can be realigned

Test Plan: Added a test and works in editor

Reviewers: zarko, schroederc

Reviewed By: schroederc

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1783