Go indexer: Fix a field-initializer ordering bug.
Summary:
When decorating the initializers of key-value fields in composite literals, the
code was not correctly checking for fields out of declaration order. Add some
tests that omit fields and permute the declaration order, and verify that the
now fixed code correctly handles these.

Note that the solution in this case involves a linear scan over the field
names, but most struct types are small enough that I don't expect this to
matter.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1803