Implement path mapping in go languageserver
Summary:
This allows us to use the settings file format from the previous kythe language server
complete with fall-through mapping. The behavior should be the same as before,
matching the first mapping that a given path or URI satisfies.
I left the xref server settings out of the settings object on purpose
because constructing the service on config file load requires the main server library
to know which service to construct. By requiring the service be externally constructed,
it's much easier to write a binary using other protocols. For this reason, host and port
are still configured via command line flags

Test Plan: Added a bunch of tests

Reviewers: zarko, schroederc

Reviewed By: schroederc

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1832