See through LinkageSpecDecls when naming nodes
Summary:
In order to be compatible with both C and C++, some headers
do the following:

```
#ifdef __cplusplus
extern "C" {
#endif
```

The C++ compiler will see the `extern "C" {` directive
(a LinkageSpecDecl), while the C compiler will not. When generating
node IDs, we were naming LinkageSpecDecls after their AST index
(the default behavior when no name is otherwise constructible).
Here, we *must* ignore them, since we want to identify declarations
in C-mode with the same declarations in C++-mode.

Reviewers: shahms, fromberger, #core_team

Reviewed By: fromberger, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1835