Decorate references inside macros.
Summary:
The indexer gives up too easily when indexing types
it's seen before, but which have macro-based locations
that can be mapped back to physical source files.
Check for this specific condition (we're emitting ranges
for types; the range we get in is valid, but in a macro;
that range is zero-length; we can follow it back to a file)
and avoid dropping out early.

Reviewers: fromberger, #core_team, shahms

Reviewed By: fromberger, #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1838