typescript: handle type imports
Summary:
TypeScript symbols exist in separate namespaces of types and values;
see
  https://www.typescriptlang.org/docs/handbook/declaration-merging.html
for a picture and testdata/type.ts for a test that verifies we model
the distinction in Kythe.

When code does
  import {x} from 'y';
we need to link 'x' to either the *type* x or the *value* x, depending
on which we imported.  (And in the case where we're importing both,
link it to the value because usually *x* is a class, which is both
a name and a value in a single location.)

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: fromberger, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1860