typescript: treat classes as both types and values
Summary:
When you write
  class Foo
you are declaring
1) a type "Foo", which is a record with the proper fields
2) a value "Foo", which is the thing referenced in expressions like
   "new Foo()"

Previously we were mixing these two into a single muddled thing.
We actually already had a test for meaning (1) in testdata/class.ts
but the test annotations were wrong (had "//", needed "//-").

This change makes us produce both the type and the value.  We currently
tag the value as a constructor which isn't quite right but's better
than it was before; I added a TODO.

Reviewers: fromberger, #core_team

Reviewed By: fromberger, #core_team

Subscribers: fromberger, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1871