Link to forward declarations of types in decorations but not in the type model
Summary:
A header that defines

```
class C;
```

may be included in a CU like

```
C* foo = nullptr;
```

where that CU never observes the complete definition of C. In this case,
we would decorate C with a reference to the tnominal node for class C.
This is non-ideal for users, as (presently) we do not generate weak links
from a class's nominal types back to any discovered definitions of those
classes. This diff instead decorates C with a link to its in-scope
declaration but does not change the type model (so the type of foo
is still a pointer to tnominal).

The old behavior is maintained for C-style declarators embedded in
declarations (`struct foo *bar`).

Reviewers: shahms, #core_team

Reviewed By: shahms, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1887