Stamp anchors used in completions.
Summary:
We're in the process of making completion a semantic
rather than a syntactic relationship. This solves a few
problems; most critically, it is currently not possible
to distinguish between multiple completion relationships
that share the exact same anchor. This can cause issues
particularly when a single macro invocation generates
a large set of definitions.

To fix this problem, we'll first generate anchors with
unique signatures for each completion head. defines* and
completes* edges will originate from these stamped anchors,
such that if `A defines B` and `A completes C`, we record
`stamp(A,B) defines B` and `stamp(A,B) completes C`.

In future diffs, we'll emit new completion edges directly between
defs and decls, using the same interface; finally, once denormalization
is ready, we can eliminate syntactic completion altogether.

Reviewers: salguarnieri, #core_team, shahms

Reviewed By: salguarnieri, #core_team, shahms

Subscribers: #core_team, schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1903