Use multiple Workspaces to support path mapping across multiple Kythe projects
Summary:
Previously, KLS used a single "PathConfig" object to control path mapping.
This made it impossible to open multiple kythe projects at once. Now the server uses
multiple `Workspace`s which perform the same task as a singular PathConfig (though could
in the future adopt more responsibility such as Workspace-local xref services). When a new
file is opened, the server will give each existing `Workspace` a chance to discover the file
(in the case of SettingsWorkspace it just looks for a path prefix). If all the workspaces file,
it falls back to a `WorkspaceProvider` function (supplied at instantiation) to produce a new
workspace.

Test Plan: Tests all pass and works in editor

Reviewers: zarko, schroederc, fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1912