Remove extraneous options from identifier service
Summary:
This commit removes the flags for inclusion of
node_kind, node_subkind, etc. in the FindReply. No consumer required them
and they could be added back in later. Furthermore, this removes MatchKind
which should have been removed earlier as per the discussion on D1807

Test Plan: It builds, no one depends on it yet.

Reviewers: zarko, schroederc, shahms

Reviewed By: shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1926