Implement identifiers service
Summary:
This change implements the identifier search service.
I've also renamed a proto type that was up to this point not depended on.
This is because the name `IdentifierMatch` described the set of all nodes
in the serving proto and a single node in the identifiers proto. This caused
a good deal of confusion when writing docs and code, so I've changed the single
node response in the identifiers proto to `Match` which at least makes them distinct.

Test Plan: There's tests! They pass! It's swell

Reviewers: schroederc, zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1927