Go indexer: Handle unmentioned imports better.
Summary:
Add a workaround to the construction rule for vnames to support names that are
only mentioned indirectly in type headers not otherwise consumed by the package
being indexed. This is not a complete solution, because unfortunately some of
the context is unavailable at this point in the process; to fully solve it we
may have to do some additional work in the extractor to detect these cases and
patch the import paths. But this will handle typical installation patterns.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1936