Invent a per-file initialization function to blame top-level calls on
Summary:
Function calls can occur outside of functions (like in static_assert,
or just in initializers). We need a place to hang these calls in the
callgraph. This diff generates a node per file that's used as the blame
site for top-level calls. The node is a function located at offset 0 of
its containing file.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2017