Always use the builtin headers built with the indexer
Summary:
If the extractor's built with a different version of Clang
than the indexer, it's very likely that the indexer will
complain about builtins (like SSE intrinsics and the like).
It's better to always use builtin files that are compatible
with the indexer's Clang. (We'll still avoid changing the
arguments list if we have cromulent header search information.)

Reviewers: shahms, fromberger, #core_team

Reviewed By: fromberger, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2029