typescript: links between module imports and exprts
Summary:
The metadata was already created on the module names in the import
statements, but we didn't have the corresponding anchor in the file
that defines the module.  To support two way linking, the metadata is
anchored on the first export keyword in a module.

The tests added cover exports on variable statements as well as stand
alone exports, as well as the case when the module is named after a
directory, but defined by an index.ts file.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2043