Include the bearer type name in concrete method signatures.
Summary:
When rendering marked source for a Go method, include the name of the type as
part of the context. The code was intended to do this, but the tests did not
actually verify this case and it wasn't working. The test now does check, and
the code now handles concrete methods as intended.

Reviewers: zarko, #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2054