Avoid emitting duplicate anchors.
Summary:
Prior to this change, the Go indexer could emit duplicate anchors under certain
circumstances, viz., when a function call occurs in the initializer of a struct
field, as in:

    s := &S{T: v.Fun(...)}
               ^^^^^^^^^^

Here, the highlighted region (an *ast.CallExpr) would be anchored to provide a
ref/call edge, and then again to provide the ref/init edge for the field.

To avoid generating the duplicate anchors, keep a cache of the nodes for whom
we have already emitted anchors. This cache is relatively cheap, since node
values are all pointers.

Also: Add a failing tests for anchor duplication.  This test will fail prior to
this commit if the verifier is allowed to check for duplicates, because the
field-initializer code can emit the same anchor that a call site wants to emit.

Reviewers: shahms, #core_team

Reviewed By: shahms, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2055