typescript: ensure 'export' is always the module
Summary:
We want the first 'export' in the file to be tagged
as the anchor for the module.  To do so we need to
catch every variant of exporting, including e.g.
  export function ...
  export type ...
Rather than handle each of these, instead catch the
first exported declaration in the main visit loop.

Note: the test only covers 'export function' (to verify that this
change fixes the previous bug, where we missed functions and others).

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: fromberger, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2065