Improve naming of function and method parameters.
Summary:
Parameters have an "owner" but the code was not using that information to
decorate their names. This change fixes that, and consolidates some shared code
into a helper function.

Prior to this change, the x in func f(x int) {} would get a name like:

    pkg.x

now it will more usefully be called

    pkg.f.x

and similarly if f is a method.

Add a test for function/method parameter naming.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2067