Simplify the construction of qualified name context in marked source.
Summary:
Add additional parent relationships for name resolution.

By default, concrete methods were not being parented to their types, and
parameters were not being parented to their functions, for the purposes of
qualified name assignment.  (This was not for any good reason, we just did not
need this information before).

Ensure that these types get mapped, which allows the subsequent improvements to
the way qualified names are assigned to parameters, and simplification of the
marked source context code.

Now that the parent tree has all the relationships we need, it is no longer
necessary to have so many special cases; we can just walk up the parent chain
directly toward the package. Pull that out into a function.

Add a method parameter name test that was missing.

No functional change.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2085